fix: moe_fused_gate routed_scaling_factor double->float ABI match
Python float maps to ctypes c_float (32-bit), but the C entry point
declared double (64-bit). This ABI mismatch would cause the value to
be read incorrectly at runtime. Internal kernel functions keep double;
the implicit float->double promotion in C++ handles the conversion.

+++ csrc/kernels/moe_fused_gate.cu
@@ -537,7 +537,7 @@ void moe_fused_gate(aiter_tensor_t* input,
                     int64_t topk_group,
                     int64_t topk,
                     int64_t num_fused_shared_experts,
-                    double routed_scaling_factor,
+                    float routed_scaling_factor,
                     hipStream_t stream)
 {
     const HipDeviceGuard device_guard(input->device_id);
